+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: mviinika <mviinika@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/08 14:13:57 by mviinika          #+#    #+#             */
-/*   Updated: 2022/03/22 14:27:53 by mviinika         ###   ########.fr       */
+/*   Updated: 2022/03/22 21:55:33 by mviinika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -94,7 +94,7 @@ double	ft_pow(double num, double pow);
 int		bintodec(signed long long bin);
 int		ft_intlen(int n);
 char	*ft_itoabase(unsigned long long int num, int base, int flag);
-char	*ft_ftoa(float num, int afterpoint);
+char	*ft_ftoa(long double num, int afterpoint);
 char	*ft_ltoa(long long int n);
 char	*ft_utoa(unsigned int n);
